scripts/requirements-import.txt
# Script: import_stok_from_excel.py için
pandas>=1.3.0
openpyxl>=3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
